backend/requirements.txt
fastapi[all]
pydantic
uvicorn
sqlalchemy
supabase
razorpay
redis[async]
python-dotenv
python-jose[cryptography]
python-multipart
httpx
passlib[bcrypt]
